fix(build): move flattenDecompositionTasks out of Next route
Next.js route modules only allow route exports; exporting a helper broke
production build type validation.
