+++ apps/web/app/api/v1/estimates/[id]/decompose/apply/route.ts
@@ -3,6 +3,7 @@ import { z } from "zod";
 import { withRole, type AuthSession } from "@/lib/auth/middleware";
 import { getPool } from "@/lib/db";
 import { seedWorkOrderTasksFromCriteria } from "@/lib/work-orders/task-time";
+import { flattenDecompositionTasks } from "@/lib/estimates/flatten-decomposition";
 import { logger } from "@/lib/logger";
 
 export const dynamic = "force-dynamic";
@@ -21,46 +22,6 @@ const bodySchema = z.object({
 
 function idFromPath(req: NextRequest): string | undefined {
   return req.nextUrl.pathname.split("/").at(-3); // .../estimates/<id>/decompose/apply
-}
-
-/**
- * Flatten AI area groups into one task list.
- * When the model still returns multiple groups, prefix task labels with the
- * group title so baselines stay readable without minting many work orders.
- */
-export function flattenDecompositionTasks(
-  groups: Array<{ title: string; scope: string; tasks: Array<{ label: string; required: boolean }> }>,
-): {
-  title: string;
-  scope: string;
-  tasks: Array<{ label: string; required: boolean }>;
-} {
-  if (groups.length === 1) {
-    return {
-      title: groups[0].title,
-      scope: groups[0].scope,
-      tasks: groups[0].tasks,
-    };
-  }
-  const tasks = groups.flatMap((g) =>
-    g.tasks.map((t) => ({
-      label: t.label.toLowerCase().startsWith(g.title.toLowerCase())
-        ? t.label
-        : `${g.title} — ${t.label}`,
-      required: t.required,
-    })),
-  );
-  const scope = groups
-    .map((g) => g.scope || g.title)
-    .filter(Boolean)
-    .join("; ");
-  return {
-    title: groups[0].title.includes(" / ")
-      ? groups[0].title
-      : groups.map((g) => g.title).join(" / ").slice(0, 200),
-    scope: scope.slice(0, 2000),
-    tasks,
-  };
 }
 
 /**
